Refactor Test struct and utilities into test scope
