+++ gateway/mw_ip_whitelist_test.go
@@ -29,7 +29,7 @@ func TestIPMiddlewarePass(t *testing.T) {
 
 	for ti, tc := range testWhiteListIPData {
 		rec := httptest.NewRecorder()
-		req := TestReq(t, "GET", "/", nil)
+		req := testReq(t, "GET", "/", nil)
 		req.RemoteAddr = tc.remote
 		if tc.forwarded != "" {
 			req.Header.Set("X-Forwarded-For", tc.forwarded)
@@ -56,7 +56,7 @@ func BenchmarkIPMiddlewarePass(b *testing.B) {
 	rec := httptest.NewRecorder()
 	for i := 0; i < b.N; i++ {
 		for ti, tc := range testWhiteListIPData {
-			req := TestReq(b, "GET", "/", nil)
+			req := testReq(b, "GET", "/", nil)
 			req.RemoteAddr = tc.remote
 			if tc.forwarded != "" {
 				req.Header.Set("X-Forwarded-For", tc.forwarded)
